@@ -9,13 +9,14 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <cucumber.version>7.11.1</cucumber.version>
-    <webdrivermanager.version>5.3.2</webdrivermanager.version>
-    <selenium-java.version>4.7.0</selenium-java.version>
-    <logback.version>1.2.11</logback.version>
-    <axe-core.version>4.4.1</axe-core.version>
+    <selenium-java.version>4.28.1</selenium-java.version>
+    <cucumber.version>7.21.1</cucumber.version>
+    <maven.compiler.plugin.version>3.11.0</maven.compiler.plugin.version>
+    <logback.version>1.5.16</logback.version>
+    <java.version>20</java.version>
+    <axe-core.version>4.10.1</axe-core.version>
     <assertj.version>3.23.1</assertj.version>
-    <appium.version>8.2.1</appium.version>
+    <appium.version>9.4.0</appium.version>
   </properties>
 
   <dependencies>
@@ -47,12 +48,6 @@
       <groupId>io.cucumber</groupId>
       <artifactId>cucumber-picocontainer</artifactId>
       <version>${cucumber.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.github.bonigarcia</groupId>
-      <artifactId>webdrivermanager</artifactId>
-      <version>${webdrivermanager.version}</version>
-      <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.assertj</groupId>
@@ -90,11 +85,11 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.8.1</version>
+        <version>${maven.compiler.plugin.version}</version>
         <configuration>
           <encoding>UTF-8</encoding>
-          <source>11</source>
-          <target>11</target>
+          <source>${java.version}</source>
+          <target>${java.version}</target>
         </configuration>
       </plugin>
     </plugins>
