@@ -1,6 +1,6 @@
 package ${package}.a11y;
 
-import com.deque.html.axecore.axeargs.AxeRunOptions;
+import com.deque.html.axecore.args.AxeRunOptions;
 import com.deque.html.axecore.results.Rule;
 import com.deque.html.axecore.selenium.AxeBuilder;
 import org.openqa.selenium.WebDriver;
